Added saving capabilities for VOI.
If the anatomy used to generate a VOI is not currently saved, it will
show a dialog box and ask if it should be saved. If so, it will be saved
in the same directory as the scene file.

Removed the link between the anat pointer and the VOI. They
didn’t (and won’t) update automatically when the slider is moved.